separate price ticker for home chart
