+++ app/Http/Livewire/Concerns/GetsCurrentPrice.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Http\Livewire\Concerns;
+
+use App\Facades\Network;
+use App\Services\Cache\NetworkStatusBlockCache;
+
+trait GetsCurrentPrice
+{
+    private function getPrice(string $currency): float
+    {
+        return (new NetworkStatusBlockCache())->getPrice(Network::currency(), $currency) ?? 0.0;
+    }
+}
