+++ resources/views/livewire/home/chart.blade.php
@@ -8,28 +8,7 @@
         'flex' => Network::canBeExchanged(),
     ])>
         <div class="flex items-center whitespace-nowrap sm:flex-1 sm:justify-between">
-            <x-home.stat
-                :title="trans('pages.home.statistics.currency_price', ['currency' => Network::currency()])"
-                class="md:hidden"
-            >
-                {{ $mainValueFiat }}
-
-                @if (ExplorerNumberFormatter::hasSymbol(Settings::currency()))
-                    {{ Settings::currency() }}
-                @endif
-            </x-home.stat>
-
-            <p class="hidden items-center space-x-2 sm:space-x-3 md:inline-flex">
-                <span class="text-sm font-semibold sm:text-3xl md:text-2xl text-theme-secondary-900 dark:text-theme-secondary-200">
-                    <span>
-                        {{ $mainValueFiat }}
-
-                        @if (ExplorerNumberFormatter::hasSymbol(Settings::currency()))
-                            {{ Settings::currency() }}
-                        @endif
-                    </span>
-                </span>
-            </p>
+            <livewire:home.price-ticker />
 
             <div class="hidden items-center space-x-3 sm:flex">
                 <x-general.dropdown.dropdown
